Added a new parameter and more parameter checks.
Added a new parameter 'skip_start', which is used to discard data from the start of the closure. The amount of seconds in 'skip_start' will be added to the 'Start time' parameter given in the times_file. This will also be applied to the times that have been interactively selected.

Added more config-file parameter checks into check_config().