Fixed a bug in the old bugfix
Fixed a bug caused by the previous "correction" where the aim was to fix cases where a measurement point
recorded exactly at the time as given in "end time" in the times file was not included for fitting. The original correction was problematic when doing manual chamber measurements and simultaneously the datetimes were created artificially as it was possible to include the first point of the next closure and if the datetime of that point was on the next day, the test discarded the measurement, because the closure was longer than two hours.

Added a check to the datafile in load_data_file, which checks for bad values in the data. Bad values are set to nan.