Updated interactive plotting and data file columns
The program now calculates the means of any columns (excluding the gas concentrations) in the data file during a measurement.
Made a huge overhaul for interactive plots. Now, all the different gas concentration time series have separate axes and y-axis values are shown. Also, PAR values are plotted if they are given in the data file.